+++ test/unit/lucid.spec.js
@@ -20,6 +20,7 @@ const { Config, setupResolver } = require('@adonisjs/sink')
 
 const helpers = require('./helpers')
 const Model = require('../../src/Lucid/Model')
+const QueryBuilder = require('../../src/Lucid/QueryBuilder')
 const DatabaseManager = require('../../src/Database/Manager')
 const VanillaSerializer = require('../../src/Lucid/Serializers/Vanilla')
 
@@ -491,7 +492,7 @@ test.group('Model', (group) => {
     }
 
     User._bootIfNotBooted()
-    await ioc.use('Database').table('users').insert([{username: 'virk'}, { username: 'nikk' }])
+    await ioc.use('Database').table('users').insert([{ username: 'virk' }, { username: 'nikk' }])
     await User.query().where('username', 'virk').update({ login_at: new Date() })
     const users = await ioc.use('Database').table('users').orderBy('id', 'asc')
     assert.equal(moment(users[0].updated_at).format('YYYY-MM-DD'), moment().format('YYYY-MM-DD'))
@@ -509,7 +510,7 @@ test.group('Model', (group) => {
     }
 
     User._bootIfNotBooted()
-    await ioc.use('Database').table('users').insert([{username: 'virk'}, { username: 'nikk' }])
+    await ioc.use('Database').table('users').insert([{ username: 'virk' }, { username: 'nikk' }])
     const users = await User.query().where('username', 'virk').fetch()
     assert.equal(users.first().toObject().full_name, 'Mr. virk')
   })
@@ -522,7 +523,7 @@ test.group('Model', (group) => {
     }
 
     User._bootIfNotBooted()
-    await ioc.use('Database').table('users').insert([{username: 'virk'}, { username: 'nikk' }])
+    await ioc.use('Database').table('users').insert([{ username: 'virk' }, { username: 'nikk' }])
     const users = await User.query().where('username', 'virk').fetch()
     assert.deepEqual(Object.keys(users.first().toObject()), ['created_at'])
   })
@@ -535,7 +536,7 @@ test.group('Model', (group) => {
     }
 
     User._bootIfNotBooted()
-    await ioc.use('Database').table('users').insert([{username: 'virk'}, { username: 'nikk' }])
+    await ioc.use('Database').table('users').insert([{ username: 'virk' }, { username: 'nikk' }])
     const users = await User.query().where('username', 'virk').fetch()
     assert.deepEqual(Object.keys(users.first().toObject()), [
       'id',
@@ -1608,7 +1609,7 @@ test.group('Model', (group) => {
     }
 
     User._bootIfNotBooted()
-    await ioc.use('Database').table('users').insert([{username: 'virk'}, { username: 'nikk' }])
+    await ioc.use('Database').table('users').insert([{ username: 'virk' }, { username: 'nikk' }])
     const users = await User.query().where('username', 'virk').setVisible(['created_at', 'id']).fetch()
     assert.deepEqual(Object.keys(users.first().toObject()), ['created_at', 'id'])
   })
@@ -1657,6 +1658,26 @@ test.group('Model', (group) => {
     assert.equal(user.username, 'foo')
   })
 
+  test('create a new row when unable to find one using trx', async (assert) => {
+    class User extends Model {
+      static boot () {
+        super.boot()
+      }
+    }
+
+    User._bootIfNotBooted()
+
+    const count = await ioc.use('Database').table('users').count('* as total')
+    assert.equal(count[0].total, 0)
+
+    const trx = await ioc.use('Database').beginTransaction()
+    const user = await User.findOrCreate({ username: 'foo' }, { username: 'virk' }, trx)
+    await trx.commit()
+
+    assert.isTrue(user.$persisted)
+    assert.equal(user.username, 'virk')
+  })
+
   test('return existing row when found one', async (assert) => {
     class User extends Model {
     }
@@ -1668,6 +1689,29 @@ test.group('Model', (group) => {
 
     await ioc.use('Database').table('users').insert({ username: 'foo' })
     const user = await User.findOrCreate({ username: 'foo' })
+    assert.isTrue(user.$persisted)
+    assert.equal(user.username, 'foo')
+    assert.equal(helpers.formatQuery(usersQuery.sql), helpers.formatQuery('select * from "users" where "username" = ? limit ?'))
+  })
+
+  test('return existing row when found one using trx', async (assert) => {
+    class User extends Model {
+      static boot () {
+        super.boot()
+      }
+    }
+
+    User._bootIfNotBooted()
+
+    let usersQuery = null
+    User.onQuery((query) => (usersQuery = query))
+
+    await ioc.use('Database').table('users').insert({ username: 'foo' })
+
+    const trx = await ioc.use('Database').beginTransaction()
+    const user = await User.findOrCreate({ username: 'foo' }, { username: 'virk' }, trx)
+    await trx.commit()
+
     assert.isTrue(user.$persisted)
     assert.equal(user.username, 'foo')
     assert.equal(helpers.formatQuery(usersQuery.sql), helpers.formatQuery('select * from "users" where "username" = ? limit ?'))
@@ -1914,5 +1958,74 @@ test.group('Model', (group) => {
 
     assert.equal(query.sql, helpers.formatQuery('select * from "users" where ("age" > ?)'))
     assert.deepEqual(query.bindings, helpers.formatBindings([18]))
+  })
+
+  test('clone query builder', async (assert) => {
+    class User extends Model {
+      static scopeAdult (builder) {
+        builder.where('age', '>', 18)
+      }
+    }
+
+    User._bootIfNotBooted()
+    assert.instanceOf(User.query().clone(), QueryBuilder)
+  })
+
+  test('clone query builder must have same constraints', async (assert) => {
+    class User extends Model {
+    }
+
+    User._bootIfNotBooted()
+    const query = User.query().where('username', 'virk')
+    const query1 = query.clone()
+
+    assert.instanceOf(query1, QueryBuilder)
+    assert.equal(query1.toSQL().sql, helpers.formatQuery('select * from "users" where "username" = ?'))
+  })
+
+  test('run whereHas inside cloned query', async (assert) => {
+    class Profile extends Model {
+    }
+
+    class User extends Model {
+      profile () {
+        return this.hasOne(Profile)
+      }
+    }
+
+    User._bootIfNotBooted()
+    Profile._bootIfNotBooted()
+
+    const query = User.query().clone().where(function () {
+      this.whereHas('profile')
+    })
+
+    assert.equal(query.toSQL().sql, helpers.formatQuery('select * from "users" where (exists (select * from "profiles" where "users"."id" = "profiles"."user_id"))'))
+  })
+
+  test('run whereHas with paginate', async (assert) => {
+    class Profile extends Model {
+    }
+
+    class User extends Model {
+      profile () {
+        return this.hasOne(Profile)
+      }
+    }
+
+    User._bootIfNotBooted()
+    Profile._bootIfNotBooted()
+
+    const user = await User.create({ username: 'virk' })
+    await user.profile().create({ profile_name: 'virk' })
+
+    let userQuery = null
+    User.onQuery((query) => (userQuery = query))
+
+    await User.query().where(function () {
+      this.whereHas('profile')
+    }).paginate()
+
+    assert.equal(userQuery.sql, helpers.formatQuery('select * from "users" where (exists (select * from "profiles" where "users"."id" = "profiles"."user_id")) limit ?'))
   })
 })